OPTIM: Reduced the number of bytes in the OWID by 4.
The signature is stored as a fixed width 64 byte array removing the need to store the length of the signature. As the signing algorithm always results in a 64 signature there was no need to store the length. If a new algorithm is used in the future then the version byte will be amended and a different fixed width for the signature could be used.