OPTIM/FEAT: Adds more information to the OWID and Offer. Demo modifications to help integrators and prepare for providing guides for CMPs (User Interface Providers) and publishers.
A new version 2 of OWID used 2 additional bytes to create a creation date in minutes. This can be used to determine if the OWID is freshly created. The demo uses this feature to display a “SWAN is setup” success notice for the first 2 minutes following initial creation of the CBID. This addresses a problem when unaccompanied trialists thought that the longer progress indicator and redirections occurs for ALL web sites as they have yet to experience visiting a further publisher.

SWAN stores all data in OWIDs as byte arrays rather than base 64 representations of the byte arrays. The payload length of the OWID is now a 2 byte integer rather than a 4 byte integer. The effect of these two changes is to reduce the length of the OWID in practice when used with SWAN.

CBID, Email and Preferences are always stored and communicated as OWIDs. This ensures that the creator domain and time are associated with them. The Offer ID has been modified to use the OWID versions of these values. The demo has been modified so that the hover tooltip over the CBID, SID and preferences displays the date and creator.

The CBID is now generated only from the SWAN Operator and not the User Interface Provider (UIP). This ensures that CBIDs can be tied back to the SWAN Network. It is expected that CBIDs generated outside the SWAN Network whilst technically valid will have a lesser reputation than those with a clear audit trail associated with them. The end point for requesting a new CBID from the SWAN Operators is /swan/api/v1/create-cbid.

The SID is now generated by the SWAN Operator based on the email known only in the SWAN network. Therefore the SID creator in the OWID could be different to the Email OWID creator. Only the SWAN Network will know the creator of the Email OWID. This will enable the SWAN Operators to easily change the hash method for the email without interfering with the method of capturing and storing the original email.

The end points for retrieval of the SWAN information have been modified to the following.

/swan/api/v1/raw – returns the RAW SWAN data as JSON key value pairs. Used by User Interface Providers (UIPs) only to get the email address and preferences with the OWID removed. The other fields that the UIP needs to continue the storage transaction and return control to the publisher are also provided in this list of pairs.

/swan/api/v1/data – returns the SWAN Pairs only for the SWAN data. Used to read only operations. The CBID, SID and preferences are OWIDs. The stopped list of advert identifiers is an array of strings.

Where not related to the inclusion of parameters in URLs the base64 RawStdEncoding has been used to reduce excess characters. Where server to server HTTP requests are involved POST messages are now used. CORS has been removed from end points that should never be called from the web browser as they contain the SWAN Access Key.

End points that were not previously supporting GZIP compression now do so.